// src/api/generated/model/sprintRewardRuleDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Title
 * Description
 * OpenAPI spec version: 1.0
 */
import type { SprintRewardRuleDtoType } from './sprintRewardRuleDtoType';
import type { SprintRewardRuleRewardDto } from './sprintRewardRuleRewardDto';

export interface SprintRewardRuleDto {
  /** Unique identifier */
  id: string;
  /** Timestamp of when the record was created */
  createdAt: string;
  /** Timestamp of the last update of the record */
  updatedAt: string;
  /** ID спринта */
  sprintId: string;
  /** Тип распределения награды */
  type: SprintRewardRuleDtoType;
  /**
   * Начало диапазона мест
   * @nullable
   */
  rankFrom: number | null;
  /**
   * Конец диапазона мест
   * @nullable
   */
  rankTo: number | null;
  /**
   * Минимальный порог баллов
   * @nullable
   */
  minPoints: number | null;
  /** Награды правила */
  rewards: SprintRewardRuleRewardDto[];
}